// dashboard/src/types/state.type.ts
export type TimingDataDriver = {
	Stats?: { TimeDiffToFastest: string; TimeDifftoPositionAhead: string }[];
	TimeDiffToFastest?: string;
	TimeDiffToPositionAhead?: string;
	GapToLeader: string;
	IntervalToPositionAhead?: {
		Value: string;
		Catching: boolean;
	};
	Line: number;
	Position: string;
	ShowPosition: boolean;
	RacingNumber: string;
	Retired: boolean;
	InPit: boolean;
	PitOut: boolean;
	Stopped: boolean;
	Status: number;
	Sectors: Sector[];
	Speeds: Speeds;
	BestLapTime: PersonalBestLapTime;
	LastLapTime: I1;
	NumberOfLaps: number; // TODO check
	KnockedOut?: boolean;
	Cutoff?: boolean;
};

export type Sector = {
	Stopped: boolean;
	Value: string;
	PreviousValue?: string;
	Status: number;
	OverallFastest: boolean;
	PersonalFastest: boolean;
	Segments: {
		Status: number;
	}[];
};

export type Speeds = {
	I1: I1;
	I2: I1;
	Fl: I1;
	St: I1;
};

export type I1 = {
	Value: string;
	Status: number;
	OverallFastest: boolean;
	PersonalFastest: boolean;
};

export type TimingStatsDriver = {
	Line: number;
	RacingNumber: string;
	PersonalBestLapTime: PersonalBestLapTime;
	BestSectors: PersonalBestLapTime[];
	BestSpeeds: {
		I1: PersonalBestLapTime;
		I2: PersonalBestLapTime;
		Fl: PersonalBestLapTime;
		St: PersonalBestLapTime;
	};
};

export type PersonalBestLapTime = {
	Value: string;
	Position: number;
};
// ...
export type Position = {
	Position: PositionItem[];
};

export type PositionItem = {
	Timestamp: string;
	Entries: Positions;
};